Limit MSVC-specific #proagma use to versions up to 2012
The optimisation-control on IsDecimalDigitForRadix(), to avoid a
warning, broke clang-cl builds (unknown #pragma); so refine the #if
condition to only apply it for versions up to MSVC2012, the one
documented to need it.

Mention the significance of suppressing optimisation in the comment
explaining it.  Also fix the grammar of the last sentence.  (The "By
moving it ..."  prefix on a sentence with "the compiler" as subject
tacitly claimed the compiler was doing the moving.)