Qt patches (#147)
* Configure support for Green Hills compiler

Trust the GHS toolchain's floating-point libraries to handle double
conversions properly.

* Suppress MSVC warning C4244 in string-to-double.cc

MSVC 2015 and later warn:
C4244: 'argument': conversion from 'const uc16' to 'char', possible loss of data
in various character-handling templates, e.g. Advance.

* Limit MSVC-specific #proagma use to versions up to 2012

The optimisation-control on IsDecimalDigitForRadix(), to avoid a
warning, broke clang-cl builds (unknown #pragma); so refine the #if
condition to only apply it for versions up to MSVC2012, the one
documented to need it.

Mention the significance of suppressing optimisation in the comment
explaining it.  Also fix the grammar of the last sentence.  (The "By
moving it ..."  prefix on a sentence with "the compiler" as subject
tacitly claimed the compiler was doing the moving.)

* Suppress unused parameter warnings from compilers

Avoids breaking Qt's build, which uses -Werror=unused-parameter

Co-authored-by: Rolland Dudemaine <rolland@dudemaine.org>
Co-authored-by: Friedemann Kleint <Friedemann.Kleint@qt.io>
Co-authored-by: Andrei Golubev <andpgolubev@gmail.com>