Revert "Add _ITERATOR_DEBUG_LEVEL=2 and _DEBUG defines (#202)" (#256)
This reverts commit dfa08c464fe164e5c70e3f080e30cc977fbb0aa9.

The original change is invalid. `_ITERATOR_DEBUG_LEVEL=2` is the default
value in debug builds. And now it's impossible to override it because
it's *appended* to `CMAKE_CXX_FLAGS_DEBUG`.
I suspect that the original problem was that release double-conversion
library was used in debug builds of Qt. This is not something that
should be fixed in double-conversion.